requirements.txt
bs4
requests
python-dateutil
pandas
selenium
pytest
